requirements.txt
pypdf==4.3.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.31.0
simplerllm==0.3.1
tenacity==9.0.0
tiktoken==0.7.0